+++ packages/w3up-client/src/service.js
@@ -3,10 +3,8 @@ import { CAR, HTTP } from '@ucanto/transport'
 import * as DID from '@ipld/dag-ucan/did'
 import { receiptsEndpoint } from '@storacha/upload-client'
 
-export const accessServiceURL = new URL('https://upload.storacha.network')
-export const accessServicePrincipal = DID.parse(
-  'did:web:upload.storacha.network'
-)
+export const accessServiceURL = new URL('https://up.storacha.network')
+export const accessServicePrincipal = DID.parse('did:web:up.storacha.network')
 
 export const accessServiceConnection = client.connect({
   id: accessServicePrincipal,
@@ -14,10 +12,8 @@ export const accessServiceConnection = client.connect({
   channel: HTTP.open({ url: accessServiceURL, method: 'POST' }),
 })
 
-export const uploadServiceURL = new URL('https://upload.storacha.network')
-export const uploadServicePrincipal = DID.parse(
-  'did:web:upload.storacha.network'
-)
+export const uploadServiceURL = new URL('https://up.storacha.network')
+export const uploadServicePrincipal = DID.parse('did:web:up.storacha.network')
 
 export const uploadServiceConnection = client.connect({
   id: uploadServicePrincipal,
@@ -25,10 +21,8 @@ export const uploadServiceConnection = client.connect({
   channel: HTTP.open({ url: accessServiceURL, method: 'POST' }),
 })
 
-export const filecoinServiceURL = new URL('https://upload.storacha.network')
-export const filecoinServicePrincipal = DID.parse(
-  'did:web:upload.storacha.network'
-)
+export const filecoinServiceURL = new URL('https://up.storacha.network')
+export const filecoinServicePrincipal = DID.parse('did:web:up.storacha.network')
 
 export const filecoinServiceConnection = client.connect({
   id: filecoinServicePrincipal,