+++ packages/upload-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [1.3.4](https://github.com/storacha/upload-service/compare/upload-api-v1.3.3...upload-api-v1.3.4) (2025-02-11)
+
+
+### Fixes
+
+* store w3s accept invocation ([#142](https://github.com/storacha/upload-service/issues/142)) ([e0c509d](https://github.com/storacha/upload-service/commit/e0c509dec75386e3275abaca1e94c4de0ccefacd))
+* use up.storacha.network ([#144](https://github.com/storacha/upload-service/issues/144)) ([ac6b5b4](https://github.com/storacha/upload-service/commit/ac6b5b4b9881f9889e99e18b38fbfb302b4fb3b5))
+
+## [1.3.3](https://github.com/storacha/upload-service/compare/upload-api-v1.3.2...upload-api-v1.3.3) (2025-01-29)
+
+
+### Other Changes
+
+* **upload-api:** add comment for clarity ([c678bde](https://github.com/storacha/upload-service/commit/c678bde352ba44e9f320f731411705b2233e83ee))
+* **upload-api:** refactor(blob-registry): pass cause as a parameter to deregister ([632348d](https://github.com/storacha/upload-service/commit/632348d1529808f67873f332c695a504ca5bd7c4))
+
 ## [1.3.2](https://github.com/storacha/upload-service/compare/upload-api-v1.3.1...upload-api-v1.3.2) (2025-01-22)